app/requirements.txt
django==2.2.*
psycopg2==2.8.*
environs==7.1.*
dj-database-url==0.5.*
gunicorn[gevent]==20.0.*
psycogreen==1.0.*
django-db-geventpool==3.2.*
celery[redis]==4.4.*
celery-redbeat==1.0.*
django-celery-results==1.1.*
django-redis==4.11.*
ddtrace==0.39.*
datadog==0.36.*
python-json-logger==0.1.11
sendgrid==6.1.*
django-sendgrid-v5==0.8.*
requests==2.22.*
black==19.10b0
isort==4.3.*
autoflake==1.3.*
mypy==0.770
djangorestframework-stubs==1.1.*
django-stubs==1.5.*
ipython==7.10.*
django-smalluuid==1.1.*
django-enumfields==1.0.*
django-phonenumber-field==4.0.*
django-timezone-field==4.*
phonenumbers==8.12.*
django_sekizai==1.0.*
django-crispy-forms==1.8.*
django-markdownify==0.8.*
django-reversion==3.0.*
markdown==3.1.*
djangorestframework==3.11.*
whitenoise==5.0.*
boto3==1.12.*
django-s3-folder-storage==0.5
django-alive==1.1.*
sentry-sdk==0.14.*
watchdog==0.9.*
pytest-django==3.8.*
django-cors-headers==3.2.1
redis==3.3.*
model-bakery==1.1.*
django-otp==0.8.*
django-formtools==2.2
django-two-factor-auth==1.12.*
qrcode==6.*
requests-mock==1.7.*
pytest-mock==2.0.*
pytest-socket==0.3.*
pytest-freezegun==0.4.*
netifaces==0.10.*
uritemplate==3.0.1
# Use version of django-nested-inline that supports read only nested formsets
git+https://github.com/gascarcella/django-nested-inline.git@read-only-change-view
cloudflare==2.6.*
reportlab==3.5.*
Pillow==7.1.*
pyyaml==5.3.*
twilio==6.39.*
jsonschema==3.2.0
pdfrw==0.4
shapely==1.7.*
django-tsvector-field==0.9.*
optimizely-sdk==3.4.*
ovrlib==0.0.3
beautifulsoup4==4.9.*
tweepy==3.8.0
selenium==3.141.*
paramiko==2.7.*
